Implemented unit tests for the CSV generation functionality in the bot.
- Created mock tests to validate the initial setup of the CSV command.
- Tested responses for scenarios with no expense records and file operation errors.
- Included checks for handling invalid income or expense types.
- Ensured proper bot messaging based on user input and data availability.
